Break up lib.rs
Brutal cut-n-paste, abusing reimpl blocks.
